- 修改细节

+++ src/main/java/com/centerm/fud_demo/service/AdminService.java
@@ -57,11 +57,11 @@ public interface AdminService {
      * 获取管理员人数
      * @return
      */
-    Integer getAdminNumber();
+    Long getAdminNumber();
 
     /**
      * 获取用户总人数
      * @return
      */
-    Integer getAllUserNumber();
+    Long getAllUserNumber();
 }
